feat: set Notify Lambda API DNS weight to 5% (#223)
* feat: set Notify Lambda API DNS weight to 5%

This will route 5% of requests to the Notify API to the API Gateway
and Lambda infrastructure.  The remaining 95% will continue to
be served by the load balancer and Kubernetes API pods.

To rollback, reset the weights and re-deploy.  The 60 second
TTL on the records will ensure a short period of impacted users.