chore: bump AWS provider to 3.x
Update the Terraform AWS provider to the 3.x release.

Going through the upgrade notes from 2.x to 3.x, the only
change thay could have impacted us was a difference in how
the `aws_route53_zone` resource returns its `name` attribute.

In 3.x, it no longer includes a `.` suffix, but we do not make
use of this attribute so there's no impact.